chore: testing ci adding special type for pages
