+++ src/__utils__/renderWithProviders.tsx
@@ -1,0 +1,22 @@
+import { render } from '@testing-library/react';
+import React, { PropsWithChildren } from 'react';
+import { Provider } from 'react-redux';
+import { setupStore } from '@/redux/store';
+
+export default function renderWithProviders(
+  ui: React.ReactElement,
+  {
+    preloadedState = {},
+    // Automatically create a store instance if no store was passed in
+    store = setupStore(preloadedState),
+    ...renderOptions
+  } = {},
+) {
+  // eslint-disable-next-line @typescript-eslint/ban-types
+  function Wrapper({ children }: PropsWithChildren<{}>): JSX.Element {
+    return <Provider store={store}>{children}</Provider>;
+  }
+
+  // Return an object with the store and all of RTL's query functions
+  return { store, ...render(ui, { wrapper: Wrapper, ...renderOptions }) };
+}
