+++ src/common/constants.ts
@@ -69,3 +69,4 @@ export const ASA_TO_ASA_PAGE_HEADER_ID = `AWAsaToAsaPageHeader`;
 export const CREATE_SWAP_BTN_ID = `AWCreateSwapButton`;
 export const ASAS_TO_ALGO_PAGE_HEADER_ID = `AWAsasToAlgoPageHeader`;
 export const MY_SWAPS_PAGE_HEADER_ID = `AWMySwapsPageHeader`;
+export const PUBLIC_SWAPS_PAGE_HEADER_ID = `AWPublicSwapsPageHeader`;
